document-intelligence-dashboard/requirements.txt
streamlit
pandas
plotly
python-docx
networkx
scikit-learn
numpy
